requirements.txt

Flask
gunicorn
mysql-connector-python
werkzeug
python-dotenv   # optional; only if you plan to use it
